// CAS-2-frontend-master/CAS-2-frontend-master/src/app/models/support/support-message.model.ts
import {constructByInterface} from '../../utils/construct-by-interface';

export interface SupportMessageInterface {
  message: string;
  filePath: string;
  url: string;
}

export class SupportMessageModel implements SupportMessageInterface {
  message: string;
  filePath: string;
  url: string;
  phone: string;

  constructor(o?: SupportMessageInterface) {
    constructByInterface(o, this);
  }
}